requirements.txt
# FastAPI and server dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
watchfiles==0.21.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# External Services
openai==1.12.0
slack-sdk==3.27.0

# CORS
starlette==0.36.3
aiofiles==23.2.1

# Date handling
pytz==2024.1
python-dateutil==2.8.2

# HTTP Client
requests==2.31.0

# Development tools
black==24.1.1
isort==5.13.2
flake8==7.0.0 